1.3.4: ObjectPool
ObjectPool is a lock- and memory-barrier-free mechanism for reusing a pool of object instances across threads, essentially by hoping to "win" data races (which, in many contexts, it commonly will).  Although it cannot guarantee a performance advantage, it is well suited to relatively small pools of somewhat-inexpensive objects where, e.g. a LinkedBlockingQueue would be too expensive relative to the object creation cost.